+++ cotengra/core.py
@@ -480,6 +480,8 @@ class ContractionTree:
         *,
         path=None,
         ssa_path=None,
+        edge_path=None,
+        optimize="auto-hq",
         autocomplete="auto",
         check=False,
         **kwargs,
@@ -499,12 +501,20 @@ class ContractionTree:
             The contraction path, a sequence of pairs of tensor ids to
             contract. The ids are linear indices into the list of temporary
             tensors, which are recycled as each contraction pops a pair and
-            appends the result. This or ``ssa_path`` must be supplied.
+            appends the result. One of ``path``, ``ssa_path`` or ``edge_path``
+            must be supplied.
         ssa_path : Sequence[Sequence[int]], optional
             The contraction path, a sequence of pairs of indices to contract.
             The indices are single use, as if the result of each contraction is
             appended to the end of the list of temporary tensors without
-            popping. This or ``path`` must be supplied.
+            popping. One of ``path``, ``ssa_path`` or ``edge_path`` must be
+            supplied.
+        edge_path : Sequence[str], optional
+            The contraction path, a sequence of indices to contract in order.
+            One of ``path``, ``ssa_path`` or ``edge_path`` must be supplied.
+        optimize : str, optional
+            If a contraction within the path contains 3 or more tensors, how to
+            optimize this subcontraction into a binary tree.
         autocomplete : "auto" or bool, optional
             Whether to automatically complete the path, i.e. contract all
             remaining nodes. If "auto" then a warning is issued if the path is
@@ -517,10 +527,17 @@ class ContractionTree:
         -------
         ContractionTree
         """
-        if int(path is None) + int(ssa_path is None) != 1:
+        if (path is None) + (ssa_path is None) + (edge_path is None) != 2:
             raise ValueError(
                 "Exactly one of ``path`` or ``ssa_path`` must be supplied."
             )
+
+        contract_opts = {"optimize": optimize, "check": check}
+
+        if edge_path is not None:
+            from .pathfinders.path_basic import edge_path_to_ssa
+
+            ssa_path = edge_path_to_ssa(edge_path, inputs)
 
         if ssa_path is not None:
             path = ssa_path
@@ -533,7 +550,7 @@ class ContractionTree:
             ssa = len(nodes)
             for p in path:
                 merge = [nodes.pop(i) for i in p]
-                nodes[ssa] = tree.contract_nodes(merge, check=check)
+                nodes[ssa] = tree.contract_nodes(merge, **contract_opts)
                 ssa += 1
             nodes = nodes.values()
         else:
@@ -541,7 +558,7 @@ class ContractionTree:
             nodes = list(tree.gen_leaves())
             for p in path:
                 merge = [nodes.pop(i) for i in sorted(p, reverse=True)]
-                nodes.append(tree.contract_nodes(merge, check=check))
+                nodes.append(tree.contract_nodes(merge, **contract_opts))
 
         if len(nodes) > 1 and autocomplete:
             if autocomplete == "auto":
@@ -552,7 +569,7 @@ class ContractionTree:
                     "Or produce an incomplete tree with `autocomplete=False`."
                 )
 
-            tree.contract_nodes(nodes, check=check)
+            tree.contract_nodes(nodes, **contract_opts)
 
         return tree
 
@@ -668,51 +685,20 @@ class ContractionTree:
         **kwargs,
     ):
         """Create a ``ContractionTree`` from an edge elimination ordering."""
-        tree = cls(inputs, output, size_dict, **kwargs)
-
-        # for efficiency track which nodes each index appears on
-        indmap = {}
-        for node in tree.gen_leaves():
-            for ix in tree.get_legs(node):
-                indmap.setdefault(ix, oset()).add(node)
-
-        for ix in edge_path:
-            # get all nodes with index
-            nodes = indmap.pop(ix)
-            if not nodes:
-                # this happens e.g. if index already
-                # contracted alongside another
-                continue
-
-            # contract the nodes
-            parent = tree.contract_nodes(
-                nodes,
-                check=check,
-                optimize=optimize,
-            )
-
-            # remove children from lookup
-            for node in nodes:
-                for jx in tree.get_legs(node):
-                    if jx != ix:
-                        indmap[jx].remove(node)
-
-            # add new parent to lookup
-            for jx in tree.get_legs(parent):
-                indmap.setdefault(jx, oset()).add(parent)
-
-        if autocomplete and len(tree.children) != tree.N - 1:
-            if autocomplete == "auto":
-                # warn that we are completing
-                warnings.warn(
-                    "Path was not complete - contracting all remaining. "
-                    "You can silence this warning with `autocomplete=True`."
-                    "Or produce an incomplete tree with `autocomplete=False`."
-                )
-
-            tree.contract_nodes(tree.children.keys(), check=check)
-
-        return tree
+        warnings.warn(
+            "ContractionTree.from_edge_path(edge_path, ...) is deprecated. Use"
+            " ContractionTree.from_path(edge_path=edge_path, ...) instead."
+        )
+        return cls.from_path(
+            inputs,
+            output,
+            size_dict,
+            edge_path=edge_path,
+            optimize=optimize,
+            autocomplete=autocomplete,
+            check=check,
+            **kwargs,
+        )
 
     def _add_node(self, node, check=False):
         if check:
@@ -1492,7 +1478,7 @@ class ContractionTree:
 
         Parameters
         ----------
-        order : None or callable, optional
+        order : None, "dfs", or callable, optional
             How to order the contractions within the tree. If a callable is
             given (which should take a node as its argument), try to contract
             nodes that minimize this function first.
@@ -2710,7 +2696,20 @@ class ContractionTree:
         )
 
     def get_path(self, order=None):
-        """Generate a standard path from the contraction tree."""
+        """Generate a standard path (with linear recycled ids) from the
+        contraction tree.
+
+        Parameters
+        ----------
+        order : None, "dfs", or callable, optional
+            How to order the contractions within the tree. If a callable is
+            given (which should take a node as its argument), try to contract
+            nodes that minimize this function first.
+
+        Returns
+        -------
+        path: tuple[tuple[int, int]]
+        """
         from bisect import bisect_left
 
         ssa = self.N
@@ -2744,7 +2743,19 @@ class ContractionTree:
         return ["einsum_path", *self.get_path(order=order)]
 
     def get_ssa_path(self, order=None):
-        """Generate a ssa path from the contraction tree."""
+        """Generate a single static assignment path from the contraction tree.
+
+        Parameters
+        ----------
+        order : None, "dfs", or callable, optional
+            How to order the contractions within the tree. If a callable is
+            given (which should take a node as its argument), try to contract
+            nodes that minimize this function first.
+
+        Returns
+        -------
+        ssa_path: tuple[tuple[int, int]]
+        """
         ssa_path = []
         pos = dict(zip(self.gen_leaves(), range(self.N)))