+++ cotengra/interface.py
@@ -125,6 +125,7 @@ def normalize_input(
     output=None,
     size_dict=None,
     shapes=None,
+    optimize=None,
     canonicalize=True,
 ):
     """Parse a contraction definition, optionally canonicalizing the indices
@@ -133,11 +134,12 @@ def normalize_input(
 
     """
     if canonicalize:
-        inputs, output, size_dict = canonicalize_inputs(
+        inputs, output, size_dict, optimize = canonicalize_inputs(
             inputs,
             output,
             shapes=shapes,
             size_dict=size_dict,
+            optimize=optimize,
         )
     elif output is None:
         # didn't canonicalize or specify output
@@ -150,7 +152,7 @@ def normalize_input(
             # didn't canonicalize and only shapes given
             size_dict = shapes_inputs_to_size_dict(shapes, inputs)
 
-    return inputs, output, size_dict
+    return inputs, output, size_dict, optimize
 
 
 _find_path_handlers = {}
@@ -159,6 +161,12 @@ _find_path_handlers = {}
 def _find_path_explicit_path(inputs, output, size_dict, optimize):
     if isinstance(optimize, list):
         optimize = tuple(optimize)
+
+    if isinstance(optimize[0], (str, int)):
+        from .pathfinders.path_basic import edge_path_to_linear
+
+        optimize = edge_path_to_linear(optimize, inputs)
+
     return optimize
 
 
@@ -268,8 +276,8 @@ def array_contract_path(
         locations should be *popped* from the list and then the result of the
         contraction *appended*.
     """
-    inputs, output, size_dict = normalize_input(
-        inputs, output, size_dict, shapes, canonicalize=canonicalize
+    inputs, output, size_dict, optimize = normalize_input(
+        inputs, output, size_dict, shapes, optimize, canonicalize
     )
 
     if cache and can_hash_optimize(optimize.__class__):
@@ -287,7 +295,14 @@ def array_contract_path(
 
 
 def _find_tree_explicit(inputs, output, size_dict, optimize):
-    return ContractionTree.from_path(inputs, output, size_dict, path=optimize)
+    if isinstance(optimize[0], (str, int)):
+        return ContractionTree.from_path(
+            inputs, output, size_dict, edge_path=optimize
+        )
+    else:
+        return ContractionTree.from_path(
+            inputs, output, size_dict, path=optimize
+        )
 
 
 def _find_tree_optimizer_search(inputs, output, size_dict, optimize, **kwargs):
@@ -409,8 +424,8 @@ def array_contract_tree(
     --------
     array_contract, array_contract_expression, einsum_tree
     """
-    inputs, output, size_dict = normalize_input(
-        inputs, output, size_dict, shapes, canonicalize=canonicalize
+    inputs, output, size_dict, optimize = normalize_input(
+        inputs, output, size_dict, shapes, optimize, canonicalize
     )
 
     nterms = len(inputs)
@@ -725,8 +740,8 @@ def array_contract_expression(
     --------
     einsum_expression, array_contract, array_contract_tree
     """
-    inputs, output, size_dict = normalize_input(
-        inputs, output, size_dict, shapes, canonicalize=canonicalize
+    inputs, output, size_dict, optimize = normalize_input(
+        inputs, output, size_dict, shapes, optimize, canonicalize
     )
 
     if constants is not None: